MINOR CHANGES: PAYMENT INTEGRATION

+++ controllers/webhook.js
@@ -3,8 +3,6 @@ import Razorpay from 'razorpay';
 import dotenv from 'dotenv';
 import fs from 'fs';
 import path from 'path';
-import { broadcast } from './sse-clients.js';
-import { clients } from './sse-clients.js';
 import cloudinary from "../config/cloudinary.js"
 import { PDFDocument } from 'pdf-lib';
 import { createHmac } from 'crypto';
@@ -51,49 +49,61 @@ export default async function RazorPayWebhook(req,res){
             }
 
             // Fill PDF
-            const pdfPath = path.resolve('vakalatnama-form.pdf');
-            const pdfBytes = fs.readFileSync(pdfPath);
-            const pdfDoc = await PDFDocument.load(pdfBytes);
-            const form = pdfDoc.getForm();
-            const pages = pdfDoc.getPages();
-            
-            const firstPage = pages[0];
-            firstPage.drawText(`Court Name`, { x: 201, y: 901, size: 12 });
-            
-            const filledPdfBytes = await pdfDoc.save();
-            const filledPdfPath = `./filled-${Date.now()}.pdf`;
-            fs.writeFileSync(filledPdfPath, filledPdfBytes);
+            if(transaction.type == "vakalatnama"){
 
-           // Upload to Cloudinary
-           const upload = await cloudinary.uploader.upload(filledPdfPath, {
-            folder: 'pdfs',
-            resource_type: 'raw'
-           });
+              const pdfPath = path.resolve('vakalatnama-form.pdf');
+              const pdfBytes = fs.readFileSync(pdfPath);
+              const pdfDoc = await PDFDocument.load(pdfBytes);
+              const form = pdfDoc.getForm();
+              const pages = pdfDoc.getPages();
+            
+              const firstPage = pages[0];
+              firstPage.drawText(`${transaction.courtName}`, { x: 201, y: 901, size: 12 });
+              firstPage.drawText(`${transaction.appealNumber}`, { x: 201, y: 881,  size: 10 });
+              firstPage.drawText(`Day`, { x: 230, y: 170,  size: 10 });
+
+              const today = new Date();
+
+              const dd = String(today.getDate()).padStart(2, '0');
+              const mm = String(today.getMonth() + 1).padStart(2, '0'); // Months are 0-indexed
+              const yy = String(today.getFullYear()).slice(-2); // Get last 2 digits
+
+              const formattedDate = `${dd}/${mm}/${yy}`;
+              firstPage.drawText(`${formattedDate}`, { x: 340, y: 170,  size: 10 });
+            
+              const filledPdfBytes = await pdfDoc.save();
+              const filledPdfPath = `./filled-${Date.now()}.pdf`;
+              fs.writeFileSync(filledPdfPath, filledPdfBytes);
+
+              // Upload to Cloudinary
+              const upload = await cloudinary.uploader.upload(filledPdfPath, {
+                folder: 'pdfs',
+                resource_type: 'raw'
+              });
            
-           fs.unlinkSync(filledPdfPath); // cleanup
+              fs.unlinkSync(filledPdfPath);
 
-          // Update transaction
-         transaction.status = true;
-         transaction.paymentId = payment.id;
-         transaction.transactionId = paymentLink.id;
-         transaction.pdf = upload.secure_url;
-         await transaction.save();
+              transaction.pdf = upload.secure_url;
+              transaction.status = true;
+              transaction.paymentId = payment.id;
+              transaction.transactionId = paymentLink.id;
 
-         console.log("PDF LINK: ", upload.secure_url)
-         console.log(transaction)
+              console.log("PDF LINK: ", upload.secure_url)
+              
+              await transaction.save();
+              console.log("PDF LINK: ", upload.secure_url)
+              console.log(transaction)
 
-         // Notify all connected clients
-         clients.forEach(c => {
-           c.write(`data: ${JSON.stringify({
-           message: 'Payment Successful',
-           pdf: transaction.pdf,
-           name: lawyer.name,
-           success: true,
-           timestamp: transaction.timestamp,
-           enrolmentNumber: transaction.enrolmentNumber
-         })}\n\n`);
-        });
-        
+            } else if(transaction.type == "membership"){
+
+                transaction.status = true;
+                transaction.paymentId = payment.id;
+                transaction.transactionId = paymentLink.id;
+              
+                await transaction.save();
+                console.log(transaction)
+            }
+
         }
 
         return res.status(200).json({ message: "success" });
